Long version id, updated some comments in the code

+++ api/src/main/java/org/apache/iceberg/view/View.java
@@ -49,7 +49,7 @@ public interface View {
    * @param versionId version ID
    * @return a version, or null if the ID cannot be found
    */
-  ViewVersion version(int versionId);
+  ViewVersion version(long versionId);
 
   /**
    * Get the version history of this table.
